LayoutTests/crypto/subtle/ecdh-import-key-derive-aes-key-expected.txt

Test deriving AES Keys with imported ECDH base key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS derivedKey.type is 'secret'
PASS derivedKey.extractable is true
PASS derivedKey.algorithm.name is 'AES-CBC'
PASS derivedKey.algorithm.length is 128
PASS derivedKey.usages is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
